requirements.txt
python-telegram-bot==20.7
selenium==4.16.0
webdriver-manager==4.0.1
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
Pillow==10.1.0
undetected-chromedriver==3.5.4 